third_party/blink/web_tests/platform/generic/css3/font-weight-multiple-selectors-expected.txt

Test that font-weight: bolder and font-weight: lighter are not cumulative when multiple selectors apply.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(outer).fontWeight is "100"
PASS getComputedStyle(bolder).fontWeight is "400"
PASS getComputedStyle(lighter).fontWeight is "100"
PASS getComputedStyle(outer).fontWeight is "200"
PASS getComputedStyle(bolder).fontWeight is "400"
PASS getComputedStyle(lighter).fontWeight is "100"
PASS getComputedStyle(outer).fontWeight is "300"
PASS getComputedStyle(bolder).fontWeight is "400"
PASS getComputedStyle(lighter).fontWeight is "100"
PASS getComputedStyle(outer).fontWeight is "400"
PASS getComputedStyle(bolder).fontWeight is "700"
PASS getComputedStyle(lighter).fontWeight is "100"
PASS getComputedStyle(outer).fontWeight is "400"
PASS getComputedStyle(bolder).fontWeight is "700"
PASS getComputedStyle(lighter).fontWeight is "100"
PASS getComputedStyle(outer).fontWeight is "500"
PASS getComputedStyle(bolder).fontWeight is "700"
PASS getComputedStyle(lighter).fontWeight is "100"
PASS getComputedStyle(outer).fontWeight is "600"
PASS getComputedStyle(bolder).fontWeight is "900"
PASS getComputedStyle(lighter).fontWeight is "400"
PASS getComputedStyle(outer).fontWeight is "700"
PASS getComputedStyle(bolder).fontWeight is "900"
PASS getComputedStyle(lighter).fontWeight is "400"
PASS getComputedStyle(outer).fontWeight is "700"
PASS getComputedStyle(bolder).fontWeight is "900"
PASS getComputedStyle(lighter).fontWeight is "400"
PASS getComputedStyle(outer).fontWeight is "800"
PASS getComputedStyle(bolder).fontWeight is "900"
PASS getComputedStyle(lighter).fontWeight is "700"
PASS getComputedStyle(outer).fontWeight is "900"
PASS getComputedStyle(bolder).fontWeight is "900"
PASS getComputedStyle(lighter).fontWeight is "700"
PASS successfullyParsed is true

TEST COMPLETE
